refactor + version selector
