+++ apps/chat/.env.development
@@ -125,3 +125,4 @@ NEXT_PUBLIC_OVERLAY_HOST="localhost:3000"
 
 ## Marketplace
 TOPICS="Business,Development,User Experience,Analysis,SQL,SDLC,Talk-To-Your-Data,RAG,Text Generation,Image Generation,Image Recognition"
+CODE_EDITOR_PYTHON_VERSIONS="python3.11,python3.12"
