+++ apps/chat/src/components/Common/ApplicationWizard/CodeAppView/CodeAppView.tsx
@@ -48,6 +48,7 @@ import { withLabel } from '@/src/components/Common/Forms/Label';
 import { MultipleComboBox } from '@/src/components/Common/MultipleComboBox';
 import { CustomLogoSelect } from '@/src/components/Settings/CustomLogoSelect';
 
+import { CodeEditor } from '../CodeEditor';
 import { ViewProps } from '../view-props';
 
 const features = [
@@ -96,6 +97,9 @@ export const CodeAppView: React.FC<ViewProps> = ({
   const files = useAppSelector(FilesSelectors.selectFiles);
   const topics = useAppSelector(SettingsSelectors.selectTopics);
   const models = useAppSelector(ModelsSelectors.selectModels);
+  const pythonVersions = useAppSelector(
+    SettingsSelectors.selectCodeEditorPythonVersions,
+  );
 
   const modelsWithFolderId = models.map((model) => ({
     ...model,
@@ -112,8 +116,13 @@ export const CodeAppView: React.FC<ViewProps> = ({
     clearErrors,
     handleSubmit: submitWrapper,
     setValue,
+    watch,
   } = useForm<FormData>({
-    defaultValues: getDefaultValues(selectedApplication, modelsWithFolderId),
+    defaultValues: getDefaultValues({
+      app: selectedApplication,
+      models: modelsWithFolderId,
+      runtime: pythonVersions[0],
+    }),
     mode: 'onChange',
     reValidateMode: 'onChange',
   });
@@ -156,6 +165,7 @@ export const CodeAppView: React.FC<ViewProps> = ({
   };
 
   register('sourceFiles', validators['sourceFiles']);
+  const sources = watch('sources');
 
   return (
     <form
@@ -267,8 +277,15 @@ export const CodeAppView: React.FC<ViewProps> = ({
           label={t('Select folder with source files')}
           rules={validators['sources']}
           error={errors.sources?.message || errors.sourceFiles?.message}
-          setValue={setValue}
         />
+
+        {sources && (
+          <CodeEditor
+            sourcesFolderId={sources}
+            setValue={setValue}
+            selectedRuntime={watch('runtime')}
+          />
+        )}
 
         <MappingsForm
           label={t('Endpoints')}
